[修改] 將呼叫者與喚起目標的權限移動至證書呼叫範圍 1. 減少檢查次數 2. 明確動作範圍
